Fix minGW compilation issues

+++ platform/windows/platform.h
@@ -21,7 +21,7 @@
 #define OG_memcpy MMXCopyMemory
 
 #define VARARGDECL(t) t _cdecl
-typedef int (__stdcall *ExtFn)();
+typedef long long int (__stdcall *ExtFn)();
 
 #ifdef _MSC_VER
 typedef __int64  FxI64;
